+++ network/gateway/joiner_bootstrap.go
@@ -51,20 +51,12 @@
 package gateway
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"sort"
 	"time"
-
-	"github.com/pkg/errors"
 
 	"github.com/insolar/insolar/insolar"
 	"github.com/insolar/insolar/insolar/pulse"
 	"github.com/insolar/insolar/instrumentation/inslogger"
-	"github.com/insolar/insolar/network"
-	"github.com/insolar/insolar/network/hostnetwork/host"
-	"github.com/insolar/insolar/network/hostnetwork/packet"
 )
 
 func newJoinerBootstrap(b *Base) *JoinerBootstrap {
@@ -78,7 +70,8 @@ type JoinerBootstrap struct {
 
 func (g *JoinerBootstrap) Run(ctx context.Context, p insolar.Pulse) {
 	logger := inslogger.FromContext(ctx)
-	permit, err := g.authorize(ctx)
+	cert := g.CertificateManager.GetCertificate()
+	permit, err := g.BootstrapRequester.Authorize(ctx, cert)
 	if err != nil {
 		logger.Warn("Failed to authorize: ", err.Error())
 		g.Gatewayer.SwitchState(ctx, insolar.NoNetworkState, p)
@@ -104,68 +97,4 @@ func (g *JoinerBootstrap) Run(ctx context.Context, p insolar.Pulse) {
 
 func (g *JoinerBootstrap) GetState() insolar.NetworkState {
 	return insolar.JoinerBootstrap
-}
-
-func (g *JoinerBootstrap) authorize(ctx context.Context) (*packet.Permit, error) {
-	logger := inslogger.FromContext(ctx)
-	cert := g.CertificateManager.GetCertificate()
-	discoveryNodes := network.ExcludeOrigin(cert.GetDiscoveryNodes(), g.NodeKeeper.GetOrigin().ID())
-
-	entropy := make([]byte, insolar.RecordRefSize)
-	if _, err := rand.Read(entropy); err != nil {
-		panic("Failed to get bootstrap entropy")
-	}
-
-	sort.Slice(discoveryNodes, func(i, j int) bool {
-		return bytes.Compare(
-			xor(*discoveryNodes[i].GetNodeRef(), entropy),
-			xor(*discoveryNodes[j].GetNodeRef(), entropy)) < 0
-	})
-
-	bestResult := &packet.AuthorizeResponse{}
-
-	for _, n := range discoveryNodes {
-		h, err := host.NewHostN(n.GetHost(), *n.GetNodeRef())
-		if err != nil {
-			logger.Warnf("Error authorizing to mallformed host %s[%s]: %s",
-				n.GetHost(), *n.GetNodeRef(), err.Error())
-			continue
-		}
-
-		logger.Infof("Trying to authorize to node: %s", h.String())
-		res, err := g.BootstrapRequester.Authorize(ctx, h, cert)
-		if err != nil {
-			logger.Warnf("Error authorizing to host %s: %s", h.String(), err.Error())
-			continue
-		}
-
-		if int(res.DiscoveryCount) < cert.GetMajorityRule() {
-			logger.Infof(
-				"Check MajorityRule failed on authorize, expect %d, got %d",
-				cert.GetMajorityRule(),
-				res.DiscoveryCount,
-			)
-
-			if res.DiscoveryCount > bestResult.DiscoveryCount {
-				bestResult = res
-			}
-
-			continue
-		}
-
-		return res.Permit, nil
-	}
-
-	if network.OriginIsDiscovery(cert) && bestResult.Permit != nil {
-		return bestResult.Permit, nil
-	}
-
-	return nil, errors.New("failed to authorize to any discovery node")
-}
-
-func xor(ref insolar.Reference, entropy []byte) []byte {
-	for i, d := range ref {
-		ref[i] = entropy[i] ^ d
-	}
-	return ref[:]
 }
